VULCAN-19 Fix packages
wiring-osgi Managers to wiring.osgi.manager
wiring-osgi Models to wiring.osgi.model
wiring-osgi Util to wiring.osgi.util
