+++ modules/apps/foundation/vulcan/vulcan-jaxrs-writer-json/src/main/java/com/liferay/vulcan/jaxrs/writer/json/internal/PageMessageBodyWriter.java
@@ -26,10 +26,10 @@ import com.liferay.vulcan.message.json.PageMessageMapper;
 import com.liferay.vulcan.pagination.Page;
 import com.liferay.vulcan.response.control.Embedded;
 import com.liferay.vulcan.response.control.Fields;
-import com.liferay.vulcan.wiring.osgi.ProviderManager;
-import com.liferay.vulcan.wiring.osgi.RelatedCollection;
-import com.liferay.vulcan.wiring.osgi.RelatedModel;
-import com.liferay.vulcan.wiring.osgi.ResourceManager;
+import com.liferay.vulcan.wiring.osgi.manager.ProviderManager;
+import com.liferay.vulcan.wiring.osgi.manager.ResourceManager;
+import com.liferay.vulcan.wiring.osgi.model.RelatedCollection;
+import com.liferay.vulcan.wiring.osgi.model.RelatedModel;
 
 import java.io.IOException;
 import java.io.OutputStream;
