+++ modules/apps/foundation/vulcan/vulcan-wiring-osgi/bnd.bnd
@@ -1,4 +1,7 @@
 Bundle-Name: Liferay Vulcan Wiring OSGi
 Bundle-SymbolicName: com.liferay.vulcan.wiring.osgi
 Bundle-Version: 1.0.0
-Export-Package: com.liferay.vulcan.wiring.osgi
+Export-Package:\
+	com.liferay.vulcan.wiring.osgi.manager,\
+	com.liferay.vulcan.wiring.osgi.model,\
+	com.liferay.vulcan.wiring.osgi.util
