+++ modules/apps/foundation/vulcan/vulcan-wiring-osgi/src/main/java/com/liferay/vulcan/wiring/osgi/internal/resource/builder/RepresentorBuilderImpl.java
@@ -17,8 +17,8 @@ package com.liferay.vulcan.wiring.osgi.internal.resource.builder;
 import com.liferay.vulcan.filter.QueryParamFilterType;
 import com.liferay.vulcan.function.TriConsumer;
 import com.liferay.vulcan.resource.builder.RepresentorBuilder;
-import com.liferay.vulcan.wiring.osgi.RelatedCollection;
-import com.liferay.vulcan.wiring.osgi.RelatedModel;
+import com.liferay.vulcan.wiring.osgi.model.RelatedCollection;
+import com.liferay.vulcan.wiring.osgi.model.RelatedModel;
 
 import java.util.List;
 import java.util.Map;
