+++ modules/apps/foundation/vulcan/vulcan-wiring-osgi/src/main/java/com/liferay/vulcan/wiring/osgi/model/RelatedCollection.java
@@ -12,27 +12,36 @@
  * details.
  */
 
-package com.liferay.vulcan.wiring.osgi;
+package com.liferay.vulcan.wiring.osgi.model;
 
-import java.util.Optional;
+import com.liferay.vulcan.filter.QueryParamFilterType;
+
 import java.util.function.Function;
 
 /**
- * Represents the relation between two models.
+ * Represents the relation between a thing and a collection.
  *
  * @author Alejandro Hernández
- * @author Carlos Sierra Andrés
- * @author Jorge Ferrer
  */
-public class RelatedModel<T, S> {
+public class RelatedCollection<T, S> {
 
-	public RelatedModel(
+	public RelatedCollection(
 		String key, Class<S> modelClass,
-		Function<T, Optional<S>> modelFunction) {
+		Function<T, QueryParamFilterType> filterFunction) {
 
 		_key = key;
 		_modelClass = modelClass;
-		_modelFunction = modelFunction;
+		_filterFunction = filterFunction;
+	}
+
+	/**
+	 * Returns the function that can be used to create the filter for the
+	 * related collection.
+	 *
+	 * @return function to calculate the related collection's filter.
+	 */
+	public Function<T, QueryParamFilterType> getFilterFunction() {
+		return _filterFunction;
 	}
 
 	/**
@@ -45,26 +54,16 @@ public class RelatedModel<T, S> {
 	}
 
 	/**
-	 * Returns the class of the related model.
+	 * Returns the class of the collection's related models.
 	 *
-	 * @return class of the related model.
+	 * @return class of the collection's related models.
 	 */
 	public Class<S> getModelClass() {
 		return _modelClass;
 	}
 
-	/**
-	 * Returns the function that can be used to retrieve the related model. It
-	 * needs a valid instance of the actual model.
-	 *
-	 * @return function to calculate the related model.
-	 */
-	public Function<T, Optional<S>> getModelFunction() {
-		return _modelFunction;
-	}
-
+	private final Function<T, QueryParamFilterType> _filterFunction;
 	private final String _key;
 	private final Class<S> _modelClass;
-	private final Function<T, Optional<S>> _modelFunction;
 
 }
