+++ modules/apps/foundation/vulcan/vulcan-wiring-osgi/src/main/java/com/liferay/vulcan/wiring/osgi/util/GenericUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.vulcan.wiring.osgi;
+package com.liferay.vulcan.wiring.osgi.util;
 
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.ReflectionUtil;
